Aawn is a prototype for Arabic mobile application to help Arab autistic children
This is a designs and implements a prototype for Arabic mobile application, called Aawn, to help and improve Arab autistic children, using their own language, in communication and emotions, as well as educational and organizational skills. Aawn is integrated with various supporting technologies based on Picture Exchange Communication System (PECS) and augmented with graphical features. The system is built and developed using Android Studio and various Cloud-based tools. The system can be extended by using different Artificial Intelligence (AI) technologies.